requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
requests>=2.31.0
python-dateutil>=2.8.2
pytz>=2023.3
streamlit>=1.29.0
matplotlib>=3.7.0
joblib>=1.3.0

#DB
SQLAlchemy>=2.0.0
pymysql>=1.1.0          
python-dotenv>=1.0.0    
